Strict cors origins + various bug fixes

+++ routes/tasks.js
@@ -18,15 +18,31 @@ router.get('/', auth, async (req, res) => {
 // @route   POST api/tasks
 // @desc    Add a new task
 router.post('/', auth, async (req, res) => {
-  const { title, description } = req.body;
+  const { 
+    title,
+    // description - Non-utilise pour le moment dans le frontend.
+  } = req.body;
 
   // Un utilisateur pourrait injecter du HTML ou du script dans la description.
+  // Ceci n'est pas un probleme pour le moment, sachant que React escape le HTML/Script par defaut.
+  // Cela devient un problem uniquement a l'usage de dangerouslySetInnerHTML, mais ce n'est jamais le cas en frontend.
+  // Un solution a ete mise en place tout de meme pour prevenir la soumission de HTML/Script dans le titre.
   try {
     const newTask = new Task({
       title,
-      description,
+      // description - Non-utilise pour le moment dans le frontend.
       user: req.user.id,
     });
+
+    // Prevenir la soumission d'un titre vide.
+    if (!title || title.trim() === "") {
+      return res.status(400).json({ msg: 'Title is required' });
+    }
+
+    // Prevenir la soumission de HTML/Script quand meme.
+    if (/<[a-z][\s\S]*>/i.test(title)) {
+      return res.status(400).json({ msg: 'HTML/Script is not allowed in the title' });
+    }
 
     const task = await newTask.save();
     res.json(task);
@@ -38,14 +54,22 @@ router.post('/', auth, async (req, res) => {
 
 // @route   PUT api/tasks/:id
 // @desc    Update a task
-// N'importe quel utilisateur authentifié peut modifier la tâche de n'importe qui d'autre s'il connaît l'ID de la tâche.
-// Il manque une vérification pour s'assurer que la tâche appartient bien à l'utilisateur qui fait la requête.
 router.put('/:id', auth, async (req, res) => {
   const { title, description, isCompleted } = req.body;
   
   try {
     let task = await Task.findById(req.params.id);
     if (!task) return res.status(404).json({ msg: 'Task not found' });
+
+    const currentUserId = req.user.id; // ID de l'utilisateur actuel
+    // Vérification pour s'assurer que la tâche appartient bien à l'utilisateur qui fait la requête.
+    if(currentUserId !== task.user.toString()) {
+      return res.status(401).json({ msg: 'You do not own this task' });
+    }
+
+    if (!title || title.trim() === "") {
+      return res.status(400).json({ msg: 'Title cannot be empty' });
+    }
 
     task = await Task.findByIdAndUpdate(req.params.id, { $set: { title, description, isCompleted } }, { new: true });
     
@@ -64,6 +88,12 @@ router.delete('/:id', auth, async (req, res) => {
   try {
     let task = await Task.findById(req.params.id);
     if (!task) return res.status(404).json({ msg: 'Task not found' });
+
+    const currentUserId = req.user.id; // ID de l'utilisateur actuel
+    // Vérification pour s'assurer que la tâche appartient bien à l'utilisateur qui fait la requête.
+    if(currentUserId !== task.user.toString()) {
+      return res.status(401).json({ msg: 'You do not own this task' });
+    }
 
     await Task.findByIdAndRemove(req.params.id);
